Split test-snippets into lib and Chimney-specific part
